% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parms}
\alias{parms}
\title{Model Parameters}
\format{
\subsection{\code{parms}}{

A list in the right format for passing to \code{runmodel}
}
}
\source{
Created by code inst/extdata/parameters.R
}
\usage{
parms
}
\description{
A default set of parameter values for running the model
}
\keyword{datasets}
