% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodel.R
\name{runmodel}
\alias{runmodel}
\title{runmodel}
\usage{
runmodel(times, p)
}
\arguments{
\item{times}{vector of times to run the model over}

\item{p}{parameter list}
}
\value{
a matrix of timeseries
}
\description{
A Simple Wrapper to Run Model
}
\details{
content for details here
}
\author{
Pete Dodd
}
